refactor: 로또 순위 결정 및 분석 결과 로직 Controller로 분리
- 발행된 로또 사전 정보를 바탕으로 사용자의 로또 순위 결정
- 로또 순위에 따라 로또 결과 분석
- 로또 결과 출력
